refactor (comment): dto 패키지 분리
