+++ src/main/java/org/sopt/domain/comment/application/service/CommentService.java
@@ -3,9 +3,9 @@ package org.sopt.domain.comment.application.service;
 import org.sopt.domain.article.domain.entity.Article;
 import org.sopt.domain.article.domain.repository.ArticleRepository;
 import org.sopt.domain.article.exception.ArticleException;
-import org.sopt.domain.comment.application.dto.CommentCreateRequest;
-import org.sopt.domain.comment.application.dto.CommentResponse;
-import org.sopt.domain.comment.application.dto.CommentUpdateRequest;
+import org.sopt.domain.comment.application.dto.request.CommentCreateRequest;
+import org.sopt.domain.comment.application.dto.response.CommentResponse;
+import org.sopt.domain.comment.application.dto.request.CommentUpdateRequest;
 import org.sopt.domain.comment.domain.entity.Comment;
 import org.sopt.domain.comment.domain.repository.CommentRepository;
 import org.sopt.domain.comment.domain.service.CommentValidator;
